# src/doc/swagger_doc.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My Book List RestAPI",
    "description": "Livros e Listas para todos os gostos!"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/review/create": {
      "post": {
        "tags": [
          "review"
        ],
        "description": "Criar nova review de um livro",
        "parameters": [
          {
            "name": "create",
            "required": true,
            "in": "body",
            "description": "Id do usuário a fazer a review livro.",
            "schema": {
              "type": "object",
              "properties": {
                "idUser": {
                  "type": "number",
                  "example": 0
                },
                "idBook": {
                  "type": "number",
                  "example": 0
                },
                "score": {
                  "type": "number",
                  "example": 10
                },
                "note": {
                  "type": "string",
                  "example": "Nota"
                },
                "status": {
                  "type": "number",
                  "example": 1
                },
                "dateOfReview": {
                  "type": "string",
                  "example": "yyyy-MM-dd 00:00:00"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "idUser": {
                      "type": "number",
                      "example": 0
                    },
                    "idBook": {
                      "type": "number",
                      "example": 0
                    },
                    "score": {
                      "type": "number",
                      "example": 10
                    },
                    "note": {
                      "type": "string",
                      "example": "Nota"
                    },
                    "status": {
                      "type": "number",
                      "example": 1
                    },
                    "dateOfReview": {
                      "type": "string",
                      "example": "yyyy-MM-dd 00:00:00"
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/review/getReviewsByIdUser": {
      "get": {
        "tags": [
          "review"
        ],
        "description": "Listar as reviews do usuário.",
        "parameters": [
          {
            "name": "idUser",
            "required": true,
            "description": "Id do usuário a ter suas reviews listadas.",
            "type": "integer",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idUser": {
                        "type": "number",
                        "example": 0
                      },
                      "idBook": {
                        "type": "number",
                        "example": 0
                      },
                      "score": {
                        "type": "number",
                        "example": 10
                      },
                      "note": {
                        "type": "string",
                        "example": "Nota"
                      },
                      "status": {
                        "type": "number",
                        "example": 1
                      },
                      "dateOfReview": {
                        "type": "string",
                        "example": "yyyy-MM-dd 00:00:00"
                      },
                      "Book": {
                        "type": "object",
                        "properties": {
                          "idBook": {
                            "type": "number",
                            "example": 0
                          },
                          "status": {
                            "type": "number",
                            "example": 0
                          },
                          "photo": {
                            "type": "string",
                            "example": "base64 ou null"
                          },
                          "name": {
                            "type": "string",
                            "example": "nome"
                          },
                          "author": {
                            "type": "string",
                            "example": "autor"
                          },
                          "publisher": {
                            "type": "string",
                            "example": "editora"
                          },
                          "dateOfPublication": {
                            "type": "string",
                            "example": "yyyy-MM-dd"
                          },
                          "synopsis": {
                            "type": "string",
                            "example": "sinopse"
                          }
                        }
                      }
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/review/delete": {
      "delete": {
        "tags": [
          "review"
        ],
        "description": "Remoção de uma review de um livro.",
        "parameters": [
          {
            "name": "idUser",
            "description": "Id do usuário a ter a review removida.",
            "type": "integer",
            "required": true,
            "in": "query"
          },
          {
            "name": "idBook",
            "description": "Id do livro a ter sua review removida.",
            "type": "integer",
            "required": true,
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Review deletada com sucesso."
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/review/getMostReviewed": {
      "get": {
        "tags": [
          "review"
        ],
        "description": "Listar os livros com maior quantidade de reviews.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idUser": {
                        "type": "number",
                        "example": 0
                      },
                      "idBook": {
                        "type": "number",
                        "example": 0
                      },
                      "score": {
                        "type": "number",
                        "example": 10
                      },
                      "note": {
                        "type": "string",
                        "example": "Nota"
                      },
                      "status": {
                        "type": "number",
                        "example": 1
                      },
                      "dateOfReview": {
                        "type": "string",
                        "example": "yyyy-MM-dd 00:00:00"
                      },
                      "Book": {
                        "type": "object",
                        "properties": {
                          "idBook": {
                            "type": "number",
                            "example": 0
                          },
                          "status": {
                            "type": "number",
                            "example": 0
                          },
                          "photo": {
                            "type": "string",
                            "example": "base64 ou null"
                          },
                          "name": {
                            "type": "string",
                            "example": "nome"
                          },
                          "author": {
                            "type": "string",
                            "example": "autor"
                          },
                          "publisher": {
                            "type": "string",
                            "example": "editora"
                          },
                          "dateOfPublication": {
                            "type": "string",
                            "example": "yyyy-MM-dd"
                          },
                          "synopsis": {
                            "type": "string",
                            "example": "sinopse"
                          }
                        }
                      }
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/review/getBestReviewed": {
      "get": {
        "tags": [
          "review"
        ],
        "description": "Listar os livros com melhor nota nas reviews.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idUser": {
                        "type": "number",
                        "example": 0
                      },
                      "idBook": {
                        "type": "number",
                        "example": 0
                      },
                      "score": {
                        "type": "number",
                        "example": 10
                      },
                      "note": {
                        "type": "string",
                        "example": "Nota"
                      },
                      "status": {
                        "type": "number",
                        "example": 1
                      },
                      "dateOfReview": {
                        "type": "string",
                        "example": "yyyy-MM-dd 00:00:00"
                      },
                      "Book": {
                        "type": "object",
                        "properties": {
                          "idBook": {
                            "type": "number",
                            "example": 0
                          },
                          "status": {
                            "type": "number",
                            "example": 0
                          },
                          "photo": {
                            "type": "string",
                            "example": "base64 ou null"
                          },
                          "name": {
                            "type": "string",
                            "example": "nome"
                          },
                          "author": {
                            "type": "string",
                            "example": "autor"
                          },
                          "publisher": {
                            "type": "string",
                            "example": "editora"
                          },
                          "dateOfPublication": {
                            "type": "string",
                            "example": "yyyy-MM-dd"
                          },
                          "synopsis": {
                            "type": "string",
                            "example": "sinopse"
                          }
                        }
                      }
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/review/dashboardByIdUser": {
      "get": {
        "tags": [
          "review"
        ],
        "description": "Buscar informaçãoes da lista de reviews do usuário pelo ID.",
        "parameters": [
          {
            "name": "idUser",
            "required": true,
            "type": "integer",
            "description": "Id do usuário a buscar informações.",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "reading": {
                      "type": "number",
                      "example": 0
                    },
                    "completed": {
                      "type": "number",
                      "example": 0
                    },
                    "stopped": {
                      "type": "number",
                      "example": 0
                    },
                    "toRead": {
                      "type": "number",
                      "example": 0
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/review/getLastReviews": {
      "get": {
        "tags": [
          "review"
        ],
        "description": "Buscar últimas reviews de um livro.",
        "parameters": [
          {
            "name": "idBook",
            "required": true,
            "type": "integer",
            "description": "Id do livro a buscar as últimas reviews.",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idUser": {
                        "type": "number",
                        "example": 0
                      },
                      "idBook": {
                        "type": "number",
                        "example": 0
                      },
                      "score": {
                        "type": "number",
                        "example": 10
                      },
                      "note": {
                        "type": "string",
                        "example": "Nota"
                      },
                      "status": {
                        "type": "number",
                        "example": 1
                      },
                      "dateOfReview": {
                        "type": "string",
                        "example": "yyyy-MM-dd 00:00:00"
                      },
                      "Book": {
                        "type": "object",
                        "properties": {
                          "idBook": {
                            "type": "number",
                            "example": 0
                          },
                          "status": {
                            "type": "number",
                            "example": 0
                          },
                          "photo": {
                            "type": "string",
                            "example": "base64 ou null"
                          },
                          "name": {
                            "type": "string",
                            "example": "nome"
                          },
                          "author": {
                            "type": "string",
                            "example": "autor"
                          },
                          "publisher": {
                            "type": "string",
                            "example": "editora"
                          },
                          "dateOfPublication": {
                            "type": "string",
                            "example": "yyyy-MM-dd"
                          },
                          "synopsis": {
                            "type": "string",
                            "example": "sinopse"
                          }
                        }
                      }
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}